pdf-to-png-conversion/requirements.txt
pdf2image>=1.16.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
reportlab>=4.0.0
